RacketOutlet_Backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
async-timeout==5.0.1
boto3==1.40.14
botocore==1.40.14
certifi==2025.8.3
charset-normalizer==3.4.3
deprecation==2.1.0
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
django-storages==1.14.6
django-unfold==0.64.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-extra-fields==3.7.0
ecdsa==0.19.1
environ==1.0
Faker==37.5.3
filetype==1.2.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jmespath==1.0.1
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
postgrest==1.1.1
psycopg==3.2.9
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
razorpay==1.4.2
realtime==2.7.0
redis==6.4.0
requests==2.32.5
s3transfer==0.13.1
sendgrid==6.12.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
weasyprint
xhtml2pdf
gunicorn